Backend/requirements.txt
pdfplumber
python-docx
spacy
fastapi
uvicorn
regex
python-multipart
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu
langchain-ollama
langchain-groq
numpy
python-dotenv
neo4j 
